+++ lib/Container/RpcExtension.php
@@ -6,9 +6,9 @@ use Phpactor\Console\Command\RpcCommand;
 use Phpactor\Rpc\HandlerRegistry;
 use Phpactor\Rpc\RequestHandler\RequestHandler;
 use Phpactor\Rpc\Handler\EchoHandler;
-use Phpactor\Rpc\Handler\StatusHandler;
+use Phpactor\Extension\Core\Rpc\StatusHandler;
 use Phpactor\Extension\WorseReflection\Rpc\GotoDefinitionHandler;
-use Phpactor\Rpc\Handler\CompleteHandler;
+use Phpactor\Extension\Completion\Rpc\CompleteHandler;
 use Phpactor\Rpc\Handler\ClassSearchHandler;
 use Phpactor\Extension\ClassMover\Rpc\ClassCopyHandler;
 use Phpactor\Extension\ClassMover\Rpc\ClassMoveHandler;
@@ -27,8 +27,8 @@ use Phpactor\Rpc\RequestHandler\ExceptionCatchingHandler;
 use Phpactor\Rpc\RequestHandler\LoggingHandler;
 use Phpactor\Rpc\Handler\NavigateHandler;
 use Phpactor\Extension\CodeTransform\Rpc\OverrideMethodHandler;
-use Phpactor\Rpc\Handler\CacheClearHandler;
-use Phpactor\Rpc\Handler\ConfigHandler;
+use Phpactor\Extension\Core\Rpc\CacheClearHandler;
+use Phpactor\Extension\Core\Rpc\ConfigHandler;
 use Phpactor\Extension\CodeTransform\Rpc\ImportClassHandler;
 use Phpactor\Extension\Extension;
 use Phpactor\Extension\ContainerBuilder;
