+++ lib/Extension/ClassMover/Application/ClassCopy.php
@@ -8,9 +8,9 @@ use Phpactor\Filesystem\Domain\Filesystem;
 use Phpactor\Phpactor;
 use Webmozart\Glob\Glob;
 use Webmozart\PathUtil\Path;
-use Phpactor\Application\Logger\ClassCopyLogger;
 use Phpactor\Filesystem\Domain\CopyReport;
 use Phpactor\Application\Helper\ClassFileNormalizer;
+use Phpactor\Extension\ClassMover\Application\Logger\ClassCopyLogger;
 
 class ClassCopy
 {
