+++ lib/Extension/ClassMover/Application/Logger/NullLogger.php
@@ -1,10 +1,12 @@
 <?php
 
-namespace Phpactor\Application\Logger;
+namespace Phpactor\Extension\ClassMover\Application\Logger;
 
 use Phpactor\Filesystem\Domain\FilePath;
 use Phpactor\ClassMover\FoundReferences;
 use Phpactor\ClassMover\Domain\Name\FullyQualifiedName;
+use Phpactor\Extension\ClassMover\Application\ClassCopyLogger;
+use Phpactor\Extension\ClassMover\Application\ClassMoverLogger;
 
 class NullLogger implements ClassCopyLogger, ClassMoverLogger
 {
