+++ lib/Extension/ClassMover/Rpc/ClassCopyHandler.php
@@ -5,7 +5,7 @@ namespace Phpactor\Extension\ClassMover\Rpc;
 use Phpactor\Extension\ClassMover\Application\ClassCopy;
 use Phpactor\Rpc\Response\OpenFileResponse;
 use Phpactor\Rpc\Response\Input\TextInput;
-use Phpactor\Application\Logger\NullLogger;
+use Phpactor\Extension\ClassMover\Application\NullLogger;
 use Phpactor\Rpc\Handler\AbstractHandler;
 
 class ClassCopyHandler extends AbstractHandler
