+++ lib/Extension/ClassToFile/ClassToFileExtension.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Phpactor\Extension\ClassToFile;
+
+use Phpactor\Extension\Extension;
+use Phpactor\Extension\ContainerBuilder;
+use Phpactor\Extension\Schema;
+
+class ClassToFileExtension implements Extension
+{
+    /**
+     * {@inheritDoc}
+     */
+    public function load(ContainerBuilder $container)
+    {
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function configure(Schema $schema)
+    {
+    }
+}
