+++ lib/Extension/Core/Command/CacheClearCommand.php
@@ -5,7 +5,7 @@ namespace Phpactor\Extension\Core\Command;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Phpactor\Application\CacheClear;
+use Phpactor\Extension\Core\Application\CacheClear;
 
 class CacheClearCommand extends Command
 {
