+++ lib/Extension/Core/Command/StatusCommand.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace Phpactor\Application;
+namespace Phpactor\Extension\Core\Command;
 
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Phpactor\Extension\Core\Application\Status;
 
 class StatusCommand extends Command
 {
