+++ lib/Extension/PathFinder/PathFinderExtension.php
@@ -7,10 +7,12 @@ use Phpactor\Extension\Extension;
 use Phpactor\Extension\ContainerBuilder;
 use Phpactor\Extension\Schema;
 use Phpactor\Extension\Container;
+use Phpactor\Extension\PathFinder\Application\Navigator;
 
 class PathFinderExtension implements Extension
 {
     const PATH_FINDER_DESTINATIONS = 'navigator.destinations';
+    const NAVIGATOR_AUTOCREATE = 'navigator.autocreate';
 
     /**
      * {@inheritDoc}
@@ -19,6 +21,14 @@ class PathFinderExtension implements Extension
     {
         $container->register('path_finder.path_finder', function (Container $container) {
             return PathFinder::fromDestinations($container->getParameter(self::PATH_FINDER_DESTINATIONS));
+        });
+
+        $container->register('application.navigator', function (Container $container) {
+            return new Navigator(
+                $container->get('path_finder.path_finder'),
+                $container->get('application.class_new'),
+                $container->getParameter(self::NAVIGATOR_AUTOCREATE)
+            );
         });
     }
 
@@ -29,6 +39,7 @@ class PathFinderExtension implements Extension
     {
         $schema->setDefaults([
             self::PATH_FINDER_DESTINATIONS => [],
+            self::NAVIGATOR_AUTOCREATE => [],
         ]);
     }
 }
