+++ lib/Extension/SourceCodeFilesystem/SourceCodeFilestem/Application/ClassSearch.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phpactor\Application;
+namespace Phpactor\Extension\SourceCodeFilesystem\SourceCodeFilestem\Application;
 
 use Phpactor\ClassFileConverter\Domain\FileToClass;
 use Phpactor\ClassFileConverter\Domain\FilePath;
