+++ tests/Unit/Application/StatusTest.php
@@ -4,7 +4,7 @@ namespace Phpactor\Tests\Unit\Application;
 
 use PHPUnit\Framework\TestCase;
 use Phpactor\Filesystem\Domain\FilesystemRegistry;
-use Phpactor\Application\Status;
+use Phpactor\Extension\Core\Application\Status;
 use Phpactor\Extension\SourceCodeFilesystem\SourceCodeFilesystemExtension;
 
 class StatusTest extends TestCase
